% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.merging.R
\name{dataset.merging}
\alias{dataset.merging}
\title{Merging all individual esets and merging them into a big eset}
\usage{
dataset.merging(
  esets,
  method = c("union", "intersect"),
  standardization = c("quantile", "robust.scaling", "scaling", "none"),
  nthread = 1
)
}
\arguments{
\item{esets}{The list containing all GSE file that need to be merged.}

\item{method}{either "unique" or "intersect" is use to for selecting geneid}

\item{standardization}{choose between "quantile", "robust.scaling", "scaling"
or "none"}

\item{nthread}{number of threads (1 by default)}
}
\value{
The merging eset
}
\description{
Merging all individual esets and merging them into a big eset
}
