% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetMerging.R
\name{datasetMerging}
\alias{datasetMerging}
\title{Merging all individual esets and merging them into a big eset}
\usage{
datasetMerging(esets, method = c("union", "intersect"),
  standardization = c("quantile", "robust.scaling", "scaling", "none"),
  nthread = 1)
}
\arguments{
\item{esets}{The list containing all GSE file that need to be merged.}

\item{method}{either "unique" or "intersect" is use to for selecting geneid}

\item{standardization}{None}

\item{nthread}{None}
}
\value{
The merging eset
}
\description{
Merging all individual esets and merging them into a big eset
}
