% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE14764.eset}
\alias{GSE14764.eset}
\title{Sample ExpressionSet from MetaGxOvarian}
\format{
A Biobase::ExpressionSet object
}
\source{
\url{http://biorxiv.org/content/biorxiv/early/2016/05/12/052910.full.pdf}
}
\usage{
GSE14764.eset
}
\description{
A Biobase::ExpressionSet from package MetaGxOvarian for the 
dataset GSE14764
}
\keyword{datasets}
