% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripWhiteSpace.R
\name{stripWhiteSpace}
\alias{stripWhiteSpace}
\title{Merging all individual esets and merging them into a big eset}
\usage{
stripWhiteSpace(str, method = c("both", "head", "tail"))
}
\arguments{
\item{str}{None}

\item{method}{None}
}
\value{
The merging eset
}
\description{
Merging all individual esets and merging them into a big eset
}
