% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.consensus.subtypes.R
\name{get.consensus.subtypes}
\alias{get.consensus.subtypes}
\title{Get consensusOV ovarian cancer subtypes}
\usage{
get.consensus.subtypes(expression.matrix, entrez.ids,
  .dataset.names.to.keep = names(esets.rescaled.classified.filteredgenes),
  remove.using.cutoff = FALSE, percentage.dataset.removed = 0.75)
}
\arguments{
\item{expression.matrix}{A matrix of gene expression values with rows as genes, columns as samples.}

\item{entrez.ids}{A vector of Entrez Gene IDs, corresponding to the rows of \code{expression.matrix}}

\item{.dataset.names.to.keep}{Names of MetaGxOvarian datasets to use for training}

\item{remove.using.cutoff}{Specify whether to classify NA for samples that do not meet a margin cutoff}

\item{percentage.dataset.removed}{If remove.using.cutoff is TRUE, then classify this percentage of
samples to NA based on margin values}
}
\value{
A list with first value \code{consensusOV.subtypes} containing a factor of subtype names;
and second value \code{rf.probs} containing a matrix of subtype probabilities
}
\description{
Get consensusOV ovarian cancer subtypes
}
\examples{
library(Biobase)
data(GSE14764.eset)
expression.matrix <- exprs(GSE14764.eset)
entrez.ids <- as.character(fData(GSE14764.eset)$EntrezGene.ID)
get.consensus.subtypes(expression.matrix, entrez.ids)
}
