% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.consensus.subtypes.R
\name{get.consensus.subtypes}
\alias{get.consensus.subtypes}
\title{Get consensusOV ovarian cancer subtypes}
\usage{
get.consensus.subtypes(expression.matrix, entrez.ids,
  .dataset.names.to.keep = names(esets.not.rescaled.classified),
  remove.using.cutoff = FALSE, percentage.dataset.removed = 0.75)
}
\arguments{
\item{expression.matrix}{A matrix of gene expression values with rows as genes, columns as samples.}

\item{entrez.ids}{A vector of Entrez Gene IDs, corresponding to the rows of \code{expression.matrix}}
}
\value{
A list with first value \code{consensusOV.subtypes} containing a factor of subtype names;
and second value \code{rf.probs} containing a matrix of subtype probabilities
}
\description{
Get consensusOV ovarian cancer subtypes
}
\examples{
data(GSE14764.eset)
expression.matrix <- GSE14764.eset
entrez.ids <- as.character(fData(GSE14764.eset)$EntrezGene.ID)
get.consensus.subtypes(expression.matrix, entrez.ids)
}
