% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.subtypes.R
\name{get.subtypes}
\alias{get.subtypes}
\title{Get ovarian cancer subtypes}
\usage{
get.subtypes(expression.dataset, entrez.ids = NULL,
  method = c("consensusOV", "Helland", "Verhaak", "Konecny", "Bentink"), ...)
}
\arguments{
\item{expression.dataset}{None}

\item{entrez.ids}{A vector of Entrez Gene IDs, corresponding to the rows of \code{expression.matrix}}

\item{method}{The subtyping method to use}

\item{...}{None
#param ... Optional parameters to be passed to the low leve function}
}
\value{
A list with first value \code{Konecny.subtypes} containing a factor of subtype names;
and second value \code{spearman.cc.vals} containing the Spearman correlation values per subtype
}
\description{
Get ovarian cancer subtypes
}
\examples{
library(Biobase)
data(GSE14764.eset)
expression.matrix <- exprs(GSE14764.eset)
entrez.ids <- as.character(fData(GSE14764.eset)$EntrezGene.ID)
get.subtypes(expression.matrix, entrez.ids, method="Konecny")
}
